requirements.txt
# Core application dependencies
numpy==2.2.4
Pillow==10.4.0
opencv-python==4.11.0.86
odfpy==1.4.1
ezodf==0.3.2
lxml==5.3.0
colorspacious==1.1.2

# Spectral analysis and plotting dependencies
matplotlib==3.9.2
pandas==2.2.3
seaborn==0.13.2
scikit-learn==1.5.2

# Excel export support
openpyxl==3.1.2

# Spreadsheet widget for Plot_3D data manager
tksheet==7.5.14

# Build dependencies
pyinstaller==6.3.0
